Register a HouseholdType with a HouseholdGrouper

Adds a HouseholdType object to a HouseholdGrouper.

Usage

addHouseholdType(object, household_type)

# S4 method for class 'HouseholdGrouper'
addHouseholdType(object, household_type)

Arguments

object

A HouseholdGrouper object.

household_type

A HouseholdType object.

Value

An updated HouseholdGrouper object.

Details

Registered household types define the household structures that may be generated during household construction.

Multiple household types can be added to the same HouseholdGrouper, allowing the synthetic population to contain a mixture of household structures.

Household types are stored internally in the household_types slot.

During execution of:

run(hg)

each registered household type is processed in turn.

Typical workflow:

Create a HouseholdGrouper.

Create one or more HouseholdType objects.

Register household types using addHouseholdType().

Execute household generation using run.

Household types are applied in the order in which they are registered.

See also

Examples

if (FALSE) { # \dontrun{

hg <- HouseholdGrouper(
  df_synth_pop = pop,
  group_by = "neighb_code"
)

hh <- HouseholdType(
  "CoupleHousehold"
)

hh <- addMembers(
  hh,
  household_position = "Parent",
  position_identifier = "adult",
  amount = 2,
  backup_position_identifiers = character()
)

hg <- addHouseholdType(
  hg,
  hh
)

result <- run(hg)

} # }