Retrieve a Household Position Definition

Returns a household-position definition stored within a HouseholdType object.

Usage

getPositionForName(object, position)

# S4 method for class 'HouseholdType'
getPositionForName(object, position)

Arguments

object

A HouseholdType object.

position

Character string identifying the required household role.

Typical values include:

"adult"

"child"

Value

A list describing the requested household-position definition.

Details

Household positions are created using addMembers and describe the composition of a household type.

Typical position identifiers include:

"adult"

"child"

The returned object contains:

position_identifier

Internal role identifier.

position

Household-position value(s) used in the synthetic population.

amount

Number of agents required for the role.

backup_position_identifiers

Alternative household-position categories that may be used if suitable agents cannot be found in the primary candidate pool.

Household-position definitions are stored internally and are referenced extensively throughout the household-generation workflow.

This method is used by:

pair_partners

group_children

matchAdultsWithChildren

An error is raised if the requested position identifier does not exist.

See also

Examples

if (FALSE) { # \dontrun{

hh <- HouseholdType(
  "Family"
)

hh <- addMembers(
  hh,
  household_position = "Parent",
  position_identifier = "adult",
  amount = 2,
  backup_position_identifiers = character()
)

adult_position <- getPositionForName(
  hh,
  "adult"
)

adult_position$amount

} # }